Pull WebRTC r8045.
BUG=81439
TEST=none
TBR=tommi

Review URL: https://codereview.chromium.org/846893004

Cr-Commit-Position: refs/heads/master@{#311179}

Former-commit-id: 7ebf2061a5af3239be36cc787a60b6a7d39b8c61

+++ content/renderer/media/rtc_video_encoder.cc
@@ -673,7 +673,8 @@ int32_t RTCVideoEncoder::Release() {
   return WEBRTC_VIDEO_CODEC_OK;
 }
 
-int32_t RTCVideoEncoder::SetChannelParameters(uint32_t packet_loss, int rtt) {
+int32_t RTCVideoEncoder::SetChannelParameters(uint32_t packet_loss,
+                                              int64_t rtt) {
   DVLOG(3) << "SetChannelParameters(): packet_loss=" << packet_loss
            << ", rtt=" << rtt;
   // Ignored.
